Add function for getting an array of profiles data
